requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap-v5==1.0.7
django-cors-headers==3.9.0
django-filter==21.1
django-heroku==0.3.1
django-microsoft-auth==2.4.1
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.1.0
python-decouple==3.5
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.8.2
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
